backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
openai==1.12.0
werkzeug==2.3.7
SQLAlchemy==2.0.23
Flask-Login==0.6.2
google-auth==2.22.0
requests==2.31.0
python-jose==3.3.0
bcrypt==4.0.1
gunicorn==21.2.0
PyJWT==2.8.0
psycopg2-binary==2.9.10
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
